9502: adding test for formatdocketentryresult [skip ci]
