+++ shared/src/persistence/elasticsearch/helpers/formatDocketEntryResult.test.js
@@ -1,0 +1,174 @@
+const { formatDocketEntryResult } = require('./formatDocketEntryResult');
+
+describe('formatDocketEntryResult', () => {
+  const emptyResults = {
+    _shards: { failed: 0, skipped: 0, successful: 1, total: 1 },
+    hits: {
+      hits: [],
+      max_score: null,
+      total: { relation: 'eq', value: 0 },
+    },
+    timed_out: false,
+    took: 66,
+  };
+
+  const matchesResults = {
+    _shards: { failed: 0, skipped: 0, successful: 1, total: 1 },
+    hits: {
+      hits: [
+        {
+          _id: 'case|102-20_case|102-20',
+          _index: 'efcms-case',
+          _score: 3.0612311,
+          _source: {
+            caseCaption: { S: 'Eve Brewer, Petitioner' },
+            contactPrimary: {
+              M: {
+                address1: { S: '67 Oak Parkway' },
+                city: { S: 'Nashville' },
+                contactId: { S: '7805d1ab-18d0-43ec-bafb-654e83405416' },
+                countryType: { S: 'domestic' },
+                email: { S: 'petitioner@example.com' },
+                name: { S: 'Eve Brewer' },
+                phone: { S: '+1 (477) 640-5087' },
+                postalCode: { S: '36042' },
+                state: { S: 'TN' },
+              },
+            },
+            contactSecondary: { M: {} },
+            docketNumber: { S: '102-20' },
+            docketNumberSuffix: { S: 'S' },
+            docketNumberWithSuffix: { S: '102-20S' },
+            irsPractitioners: {
+              L: [
+                {
+                  M: {
+                    barNumber: { S: 'RT6789' },
+                    contact: {
+                      M: {
+                        address1: { S: '234 Main St' },
+                        address2: { S: 'Apartment 4' },
+                        address3: { S: 'Under the stairs' },
+                        city: { S: 'Chicago' },
+                        countryType: { S: 'domestic' },
+                        phone: { S: '+1 (555) 555-5555' },
+                        postalCode: { S: '61234' },
+                        state: { S: 'IL' },
+                      },
+                    },
+                    email: { S: 'irsPractitioner@example.com' },
+                    name: { S: 'Test IRS Practitioner' },
+                    pk: { S: 'case|102-20' },
+                    role: { S: 'irsPractitioner' },
+                    section: { S: 'irsPractitioner' },
+                    serviceIndicator: { S: 'Electronic' },
+                    sk: {
+                      S: 'irsPractitioner|5805d1ab-18d0-43ec-bafb-654e83405416',
+                    },
+                    userId: { S: '5805d1ab-18d0-43ec-bafb-654e83405416' },
+                  },
+                },
+              ],
+            },
+            privatePractitioners: {
+              L: [
+                {
+                  M: {
+                    barNumber: { S: 'PT5432' },
+                    contact: {
+                      M: {
+                        address1: { S: '234 Main St' },
+                        address2: { S: 'Apartment 4' },
+                        address3: { S: 'Under the stairs' },
+                        city: { S: 'Chicago' },
+                        countryType: { S: 'domestic' },
+                        phone: { S: '+1 (555) 555-5555' },
+                        postalCode: { S: '61234' },
+                        state: { S: 'IL' },
+                      },
+                    },
+                    email: { S: 'privatePractitioner1@example.com' },
+                    name: { S: 'Test Private Practitioner' },
+                    pk: { S: 'case|102-20' },
+                    representingPrimary: { BOOL: true },
+                    role: { S: 'privatePractitioner' },
+                    section: { S: 'privatePractitioner' },
+                    serviceIndicator: { S: 'Electronic' },
+                    sk: {
+                      S: 'privatePractitioner|ad07b846-8933-4778-9fe2-b5d8ac8ad728',
+                    },
+                    userId: { S: 'ad07b846-8933-4778-9fe2-b5d8ac8ad728' },
+                  },
+                },
+              ],
+            },
+            receivedAt: { S: '2020-01-21T16:41:39.474Z' },
+            sealedDate: { S: '2020-01-21T16:48:28.364Z' },
+          },
+          _type: '_doc',
+        },
+      ],
+      max_score: 3.0612311,
+      total: { relation: 'eq', value: 1 },
+    },
+    timed_out: false,
+    took: 7,
+  };
+
+  it('should unmarshall the case when it is not found in the caseMap', async () => {
+    const mockCaseMap = {
+      '123-99': {},
+    };
+
+    const results = await formatDocketEntryResult({
+      caseMap: mockCaseMap,
+      hit: matchesResults,
+      sourceUnmarshalled: {},
+    });
+
+    expect(results).toMatchObject({
+      results: [
+        {
+          _score: null,
+          caseCaption: 'Brett Osborne, Petitioner',
+          docketEntryId: '25100ec6-eeeb-4e88-872f-c99fad1fe6c7',
+          docketNumber: '312-21',
+          docketNumberWithSuffix: '312-21W',
+          documentTitle: 'Order of Dismissal for Lack of Jurisdiction',
+          documentType: 'Order of Dismissal for Lack of Jurisdiction',
+          eventCode: 'ODJ',
+          filingDate: '2021-08-20T18:16:39.420Z',
+          irsPractitioners: [],
+          isCaseSealed: false,
+          isDocketEntrySealed: false,
+          isSealed: undefined,
+          isStricken: false,
+          judge: 'Carluzzo',
+          numberOfPages: 2,
+          petitioners: [
+            {
+              address1: '68 Fabien Freeway',
+              address2: 'Suscipit animi solu',
+              address3: 'Architecto assumenda',
+              city: 'Aspernatur nostrum s',
+              contactId: '7805d1ab-18d0-43ec-bafb-654e83405416',
+              contactType: 'primary',
+              countryType: 'domestic',
+              email: 'petitioner@example.com',
+              isAddressSealed: false,
+              name: 'Brett Osborne',
+              phone: '+1 (537) 235-6147',
+              postalCode: '89499',
+              sealedAndUnavailable: false,
+              serviceIndicator: 'Electronic',
+              state: 'AK',
+            },
+          ],
+          privatePractitioners: [],
+          signedJudgeName: 'Maurice B. Foley',
+        },
+      ],
+      total: 1,
+    });
+  });
+});
